[crashes] Extract generate update function
